Moved a bunch of stuff around to follow a more strict android packaging layout.
