+++ app/src/main/AndroidManifest.xml
@@ -6,13 +6,13 @@
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
 
     <application
-        android:name="edu.purdue.app.PurdueApplication"
+        android:name=".application.PurdueApplication"
         android:allowBackup="true"
         android:icon="@drawable/ic_launcher"
         android:label="@string/app_name"
         android:theme="@style/Theme.Purdue" >
         <activity
-            android:name="edu.purdue.app.main.MainMenuActivity"
+            android:name=".activity.MainMenuActivity"
             android:label="@string/app_name"
             android:screenOrientation="portrait">
             <intent-filter>
@@ -22,9 +22,9 @@
             </intent-filter>
         </activity>
         <activity
-            android:name="edu.purdue.app.WebViewActivity"
+            android:name=".activity.WebViewActivity"
             android:label="@string/title_activity_web_view"
-            android:parentActivityName="edu.purdue.app.main.MainMenuActivity" >
+            android:parentActivityName=".activity.MainMenuActivity" >
 
             <!-- The meta-data element is needed for versions lower than 4.1 -->
             <meta-data
@@ -37,9 +37,9 @@
             android:value="@integer/google_play_services_version" />
 
         <activity
-            android:name=".prefs.SettingsActivity"
+            android:name=".activity.SettingsActivity"
             android:label="@string/title_activity_settings"
-            android:parentActivityName="edu.purdue.app.main.MainMenuActivity" >
+            android:parentActivityName=".activity.MainMenuActivity" >
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
                 android:value="edu.purdue.app.main.MainMenuActivity" />
