+++ app/src/main/java/edu/purdue/app/fragment/DataSyncPreferenceFragment.java
@@ -1,4 +1,4 @@
-package edu.purdue.app.prefs;
+package edu.purdue.app.fragment;
 
 import android.annotation.TargetApi;
 import android.os.Build;
@@ -6,23 +6,23 @@ import android.os.Bundle;
 import android.preference.PreferenceFragment;
 
 import edu.purdue.app.R;
+import edu.purdue.app.activity.SettingsActivity;
 
 /**
- * This fragment shows general preferences only. It is used when the
+ * This fragment shows data and sync preferences only. It is used when the
  * activity is showing a two-pane settings UI.
  */
 @TargetApi(Build.VERSION_CODES.HONEYCOMB)
-public class GeneralPreferenceFragment extends PreferenceFragment {
+public class DataSyncPreferenceFragment extends PreferenceFragment {
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        addPreferencesFromResource(R.xml.pref_general);
+        addPreferencesFromResource(R.xml.pref_data_sync);
 
         // Bind the summaries of EditText/List/Dialog/Ringtone preferences
         // to their values. When their values change, their summaries are
         // updated to reflect the new value, per the Android Design
         // guidelines.
-        SettingsActivity.bindPreferenceSummaryToValue(findPreference("example_text"));
-        SettingsActivity.bindPreferenceSummaryToValue(findPreference("example_list"));
+        SettingsActivity.bindPreferenceSummaryToValue(findPreference("sync_frequency"));
     }
 }
