+++ app/src/main/java/edu/purdue/app/listener/CustomOnSharedPreferenceChangeListener.java
@@ -1,9 +1,11 @@
-package edu.purdue.app.prefs;
+package edu.purdue.app.listener;
 
 import android.content.Context;
 import android.content.SharedPreferences;
 
 import com.google.android.gms.analytics.GoogleAnalytics;
+
+import edu.purdue.app.constants.PrefConstants;
 
 /**
  * Created by david on 5/21/14 for Purdue
