+++ app/src/main/res/layout/activity_web_view.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="edu.purdue.app.WebViewActivity">
+    tools:context="edu.purdue.app.activity.WebViewActivity">
 
     <WebView
         android:layout_width="fill_parent"
