+++ app/src/main/res/menu/main_menu.xml
@@ -1,6 +1,6 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="edu.purdue.app.main.MainMenuActivity" >
+    tools:context="edu.purdue.app.activity.MainMenuActivity" >
 
     <item android:id="@+id/action_settings"
         android:title="@string/action_settings"
