+++ app/src/main/res/menu/web_view.xml
@@ -1,6 +1,6 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="edu.purdue.app.WebViewActivity" >
+    tools:context="edu.purdue.app.activity.WebViewActivity" >
 
 
     <item
