+++ app/src/main/res/xml/app_tracker.xml
@@ -8,10 +8,10 @@
     <bool name="ga_reportUncaughtExceptions">true</bool>
 
     <!-- The screen names that will appear in reports -->
-    <screenName name="edu.purdue.app.main.MainMenuActivity">
+    <screenName name="edu.purdue.app.activity.MainMenuActivity">
         PurdueApp MainMenuActivity
     </screenName>
-    <screenName name="edu.purdue.app.WebViewActivity">
+    <screenName name="edu.purdue.app.activity.WebViewActivity">
         PurdueApp WebViewActivity
     </screenName>
